Moved mount collection to disk collector
