+++ Vostok.Metrics.System/Host/DiskSpaceCollector.cs
@@ -5,10 +5,14 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using Vostok.Metrics.System.Helpers;
 
+// ReSharper disable InconsistentNaming
+
 namespace Vostok.Metrics.System.Host
 {
-    internal class DiskSpaceCollector
+    internal class DiskSpaceCollector : IDisposable
     {
+        private readonly ReusableFileReader mountsReader_Linux;
+        private Dictionary<string, string> mountDiskMap = new Dictionary<string, string>();
         private readonly Func<DriveInfo, bool> systemFilter;
         private readonly Func<string, string> nameFormatter;
 
@@ -23,6 +27,7 @@ namespace Vostok.Metrics.System.Host
             {
                 systemFilter = FilterDisks_Linux;
                 nameFormatter = FormatDiskName_Linux;
+                mountsReader_Linux = new ReusableFileReader("/proc/mounts");
             }
         }
 
@@ -65,12 +70,39 @@ namespace Vostok.Metrics.System.Host
             }
         }
 
-        private bool FilterDisks_Linux(DriveInfo disk) => NativeHostMetricsCollector_Linux.MountDiskMap.ContainsKey(disk.Name);
+        private void UpdateMountMap()
+        {
+            mountDiskMap = new Dictionary<string, string>();
+
+            try
+            {
+                foreach (var mountLine in mountsReader_Linux.ReadLines())
+                {
+                    if (FileParser.TrySplitLine(mountLine, 2, out var parts) && parts[0].Contains("/dev/"))
+                        mountDiskMap[parts[1]] = parts[0];
+                }
+            }
+            catch (Exception error)
+            {
+                InternalErrorLogger.Warn(error);
+            }
+        }
+
+        private bool FilterDisks_Linux(DriveInfo disk)
+        {
+            UpdateMountMap();
+            return mountDiskMap.ContainsKey(disk.Name);
+        }
 
         private bool FilterDisks_Windows(DriveInfo disk) => true;
 
         private string FormatDiskName_Windows(string diskName) => diskName.Replace(":\\", string.Empty);
 
-        private string FormatDiskName_Linux(string diskName) => NativeHostMetricsCollector_Linux.MountDiskMap[diskName].Replace("/dev/", string.Empty);
+        private string FormatDiskName_Linux(string diskName) => mountDiskMap[diskName].Replace("/dev/", string.Empty);
+
+        public void Dispose()
+        {
+            mountsReader_Linux?.Dispose();
+        }
     }
 }
