+++ Vostok.Metrics.System/Host/NativeHostMetricsCollector_Linux.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
@@ -11,12 +10,10 @@ namespace Vostok.Metrics.System.Host
 {
     internal class NativeHostMetricsCollector_Linux : IDisposable
     {
-        public static Dictionary<string, string> MountDiskMap = new Dictionary<string, string>();
         private readonly Regex pidRegex = new Regex("[0-9]+$", RegexOptions.Compiled);
         private readonly ReusableFileReader systemStatReader = new ReusableFileReader("/proc/stat");
         private readonly ReusableFileReader memoryReader = new ReusableFileReader("/proc/meminfo");
         private readonly ReusableFileReader descriptorInfoReader = new ReusableFileReader("/proc/sys/fs/file-nr");
-        private readonly ReusableFileReader mountsReader = new ReusableFileReader("/proc/mounts");
         private readonly HostCpuUtilizationCollector cpuCollector = new HostCpuUtilizationCollector();
 
         public void Dispose()
@@ -24,7 +21,6 @@ namespace Vostok.Metrics.System.Host
             systemStatReader.Dispose();
             memoryReader.Dispose();
             descriptorInfoReader.Dispose();
-            mountsReader.Dispose();
         }
 
         public void Collect(HostMetrics metrics)
@@ -55,8 +51,6 @@ namespace Vostok.Metrics.System.Host
                 metrics.ThreadCount = perfInfo.ThreadCount.Value;
                 metrics.ProcessCount = perfInfo.ProcessCount.Value;
             }
-
-            UpdateMountMap();
         }
 
         private SystemStat ReadSystemStat()
@@ -149,24 +143,6 @@ namespace Vostok.Metrics.System.Host
             }
 
             return result;
-        }
-
-        private void UpdateMountMap()
-        {
-            MountDiskMap = new Dictionary<string, string>();
-
-            try
-            {
-                foreach (var mountLine in mountsReader.ReadLines())
-                {
-                    if (FileParser.TrySplitLine(mountLine, 2, out var parts) && parts[0].Contains("/dev/"))
-                        MountDiskMap[parts[1]] = parts[0];
-                }
-            }
-            catch (Exception error)
-            {
-                InternalErrorLogger.Warn(error);
-            }
         }
 
         private class PerformanceInfo
